# Conflicts:
#	app/Console/Kernel.php
#	app/Helpers/AssetHelpers.php
#	app/Models/Notification.php
#	app/Services/RankService.php
#	config/lorekeeper/admin_sidebar.php
#	config/lorekeeper/extension_tracker.php
#	routes/lorekeeper/admin.php

+++ app/Services/RankService.php
@@ -5,6 +5,8 @@ namespace App\Services;
 use App\Models\Rank\Rank;
 use App\Models\User\User;
 use Illuminate\Support\Facades\DB;
+use App\Models\Rank\RankPower;
+use App\Models\RankReward;
 
 class RankService extends Service {
     /*
@@ -142,6 +144,9 @@ class RankService extends Service {
             if (User::where('rank_id', $rank->id)->exists()) {
                 throw new \Exception('There are currently user(s) with the selected rank. Please change their rank before deleting this one.');
             }
+
+            // aaand check it's not set as a rank reward
+            if(RankReward::where('rank_id', $rank->id)->exists()) throw new \Exception("An rank reward with this rank exists. Please change its rank or delete it first.");
 
             $rank->powers()->delete();
             $rank->delete();
